# api/openapi/openapi.yaml
openapi: 3.1.0
...
info:
  title: Baseproof Trust Network Courts API
  version: 0.1.0
  summary: Multi-court federation API for judicial log entries, artifacts, and cross-court verification.
  description: |
    The Trust Network Courts (JN) API is the HTTP surface for an Baseproof
    deployment serving one or more courts on a single listener. The
    composer mounts three sub-surfaces:

      - `/v1/judicial/*` — domain-aware judicial flows (cases, appeals,
        enforcement, parties, onboarding, artifacts, verification,
        monitoring, consortium federation, delegation, topology).
      - `/v1/entries/*`, `/v1/artifacts/*`, `/v1/delegations/*`,
        `/v1/keys/*`, `/v1/dids`, `/v1/scope/*` — generic Baseproof
        protocol surface (build/sign/submit, key custody, scope
        amendment).
      - `/v1/verify/*` — read-side verification (origin, authority,
        delegation, cross-log, fraud-proof).

    Routing is multi-tenant payload-driven: every entry's destination
    is sourced from `entry.Header.Destination`, looked up in the
    Bundle Registry. There is exactly one handler tree per route
    family, shared across all registered destinations.

    ### Build / sign / submit pattern

    Most write endpoints under `/v1/judicial/*` return a
    `BuildResponse` carrying the unsigned entry bytes and the
    signing payload. The caller signs the payload externally
    (typically through Privy or an SCW party signer) and submits the
...

    ### Authentication

...

...

...

...
servers:
...
security:
  - mTLS: []
  - bearerAuth: []

# ─────────────────────────────────────────────────────────────────────
# Paths
# ─────────────────────────────────────────────────────────────────────
paths:
...
  /v1/verify/complete/{logID}/{pos}:
    get:
      tags: [Verify]
      summary: Run the SDK Path C composite (Signatures + Authority + Origin) on the entry at the given position.
      operationId: verifyComplete
      parameters:
        - name: logID
          in: path
          required: true
          schema: { type: string }
        - name: pos
          in: path
          required: true
          schema: { type: integer, format: uint64 }
      responses:
        '200':
          description: Unified per-stage report. `all_green` is true iff every opted-in stage passed.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/VerifyResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/BadRequest' }
  /v1/verify/cross-log:
    post:
      tags: [Verify]
      summary: Verify a generic cross-log proof.
      operationId: verifyCrossLog
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CrossLogProofRequest' }
      responses:
        '200':
          description: Verdict.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CrossLogProofResult' }
        '400': { $ref: '#/components/responses/BadRequest' }
...
components:
...
  securitySchemes:
    mTLS:
      type: mutualTLS
      description: |
        Client certificate carries the caller's DID in a SAN URI.
        Composer-level middleware lifts it into request context.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT validated against a configured JWKS endpoint. The `sub`
        claim is the caller DID.

  # ── Reusable parameters ───────────────────────────────────────────
...
  responses:
...
    BadRequest:
      description: Invalid request body or missing required fields.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message.
...
    CrossLogProofRequest:
      type: object
      required: [proof, source_log_did]
      properties:
        proof:
          type: object
          additionalProperties: true
          description: Marshaled types.CrossLogProof.
        source_log_did: { type: string }
        source_witness_keys_b64:
          type: array
          items: { type: string, format: byte }
        source_witness_quorum: { type: integer }
    CrossLogProofResult:
      type: object
      required: [verified]
      properties:
        verified: { type: boolean }
        error: { type: string }
...
    VerifyResult:
      type: object
      required: [verified]
      properties:
        verified: { type: boolean }
        error: { type: string }
        details: { type: object, additionalProperties: true }
...